update to css (parser) 2.0.0
will be used in Penthouse core to produce AST object
to work on, to replace current regex parsing.

Exact 2.0.0 version is requred, because:
< 2.0.0 did not have silent flag, and would just crash on
invalid css, which is not acceptable.
> 2.0.0 doesn't work in Phantom, due to problems with the inherits library:
https://github.com/reworkcss/css/commit/5946f617d6300d7c376aa7c49d329738c0767a25

Would be great to go above 2.0.0 in order to make use of the parsingErrors field
on stylesheet node in returned AST - to inform users that their invalid css is
likely to cause incorrect critical css. Perhaps look into fork in future.