disable ensure same case after invalid css normalising
did profiling on Penthouse and when running with invalid css
saw roughly 85% of time spent in handleRuleSelectorCase
(40s!! for 1.2MB large css - causing penthouse timeouts).

Cannot even think of why the critical css needs to be faithful to original case,
given that it only matters for JS, and not for CSS, so shouldn't matter.