keep animation props, and used @keyframes
if they are declared on critical elements, they will be triggered
immediately after the first render, and are therefor needed.
Especially considering the use case where a fade-in  @keyframe animation
is responsible for fading an element in from 0 to 1 opacity, where
the default ruleset for element has opacity 1. Without keeping
animation+keyframe, the element would flicker back to 0 once
the full css loads.