cleanup generated file on disk
the ast object. the tmp library used would create it by itself
after the process exists (node, that is), but if it goes on for
a long time with multiple calls to Penthouse, those files can
start to take up some space - better to clear when we know we
don't need them anymore.