implement setTimeout workaround for js disabled page.evaluate execution
setTimeout, setInterval - does not work inside page.evaluate
when JS is disabled:
https://github.com/GoogleChrome/puppeteer/issues/562

So to keep same functionality, added this workaround.

Could also have moved the simple setTimeout outside of the page.evaluate in this case;
but it could be useful for other purposes to be able to use it inside page.evaluate.