add keepLargerMediaQueries option
given that Penthouse works very well to generate only one critical css file
per page (as opposed to different ones for different viewport sizes),
and that this is what I recommend - it makes sense to allow to keep larger
media queries so truly make the critical css work for all sizes.

keepLargerMediaQueries enables this behavior, but is defaulting to false,
to be backwards compatible.