move render waiting out of critical css generation
the purpose of this waiting has been to ensure extra time after page load,
before the critical css pruning starts. No need for this wait to happen
in the evaluate function inside of browser sandbox. It also no longer
worked in Puppeteer@1.6, as requestAnimationFrame is no longer called
(with JS disabled)